+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Bugfixes
 - [#2225](https://github.com/influxdb/influxdb/pull/2225): Make keywords completely case insensitive
+- [#2228](https://github.com/influxdb/influxdb/pull/2228): Accept keyword default unquoted in ALTER RETENTION POLICY statement
 
 ## v0.9.0-rc22 [2015-04-09]